Move error toast to button component
Also launch on enter when input is focused
