+++ src/DesktopCard.js
@@ -1,9 +1,6 @@
 import React from 'react';
 import classNames from 'classnames';
 
-import {
-} from 'flight-webapp-components';
-
 import { CardFooter } from './CardParts';
 import LaunchDesktopButton from './LaunchDesktopButton';
 import { prettyDesktopName } from './utils';
@@ -46,40 +43,11 @@ function DesktopCard({ desktop }) {
             size="sm"
             className="mr-2"
             desktop={desktop}
-            errorToast={launchErrorToast}
           />
         </div>
       </CardFooter>
     </div>
   );
-}
-
-function launchErrorToast({ clusterName, desktop, launchError }) {
-  const desktopName = prettyDesktopName[desktop.id];
-  let body = (
-    <div>
-      Unfortunately there has been a problem launching your
-      {' '}<strong>{desktopName}</strong> desktop session.  Please try
-      again and, if problems persist, help us to more quickly rectify the
-      problem by contacting us and letting us know.
-    </div>
-  );
-  if (launchError === 'Desktop Not Prepared') {
-    body = (
-      <div>
-        <strong>{desktopName}</strong> has not yet been fully configured.  If
-        you would like to use this desktop please contact the system
-        administrator for {' '}<em>{clusterName}</em> and ask them to prepare
-        this desktop.
-      </div>
-    );
-  }
-
-  return {
-    body,
-    icon: 'danger',
-    header: 'Failed to launch desktop',
-  };
 }
 
 export default DesktopCard;
