+++ src/LaunchDesktopButton.js
@@ -11,11 +11,11 @@ import {
 
 import ModalContainer from "./ModalContainer";
 import { useLaunchSession } from './api';
+import { prettyDesktopName } from './utils';
 
 function LaunchDesktopButton({
   className,
   desktop,
-  errorToast,
   children,
   color,
   size
@@ -23,7 +23,7 @@ function LaunchDesktopButton({
   const [modal, setModal] = useState(false);
   const toggle = () => setModal(!modal);
   const clusterName = useContext(ConfigContext).clusterName;
-  const modalTitle = <span>Configure this session</span>
+  const modalTitle = <span>Configure this session</span>;
   const nameRef = useRef(null);
   const { request, post } = useLaunchSession();
   const { addToast } = useToast();
@@ -34,13 +34,24 @@ function LaunchDesktopButton({
       if (request.response.ok) {
         history.push(`/sessions/${responseBody.id}`);
       } else {
-        addToast(errorToast({
+        addToast(launchErrorToast({
           clusterName: clusterName,
           desktop: desktop,
           launchError: utils.errorCode(responseBody),
         }));
       }
     });
+  };
+
+  const handleSubmit = e => {
+    launchSession();
+    toggle();
+  };
+
+  const handleKeyDown = (event) => {
+    if (event.key === 'Enter') {
+      handleSubmit();
+    }
   };
 
   const leftButton = (
@@ -56,7 +67,7 @@ function LaunchDesktopButton({
     <Button
       data-testid="session-launch-button"
       className="btn-sm ml-2"
-      onClick={() => { launchSession(); toggle(); }}
+      onClick={handleSubmit}
   >
       <i className="fa fa-bolt mr-1"></i>
       Launch
@@ -97,11 +108,40 @@ function LaunchDesktopButton({
           placeholder="Session name"
           type="text"
           ref={nameRef}
+	  onKeyDown={handleKeyDown}
           autoFocus={true}
         />
       </ModalContainer>
     </div>
   );
+}
+
+function launchErrorToast({ clusterName, desktop, launchError }) {
+  const desktopName = prettyDesktopName[desktop.id];
+  let body = (
+    <div>
+      Unfortunately there has been a problem launching your
+      {' '}<strong>{desktopName}</strong> desktop session.  Please try
+      again and, if problems persist, help us to more quickly rectify the
+      problem by contacting us and letting us know.
+    </div>
+  );
+  if (launchError === 'Desktop Not Prepared') {
+    body = (
+      <div>
+        <strong>{desktopName}</strong> has not yet been fully configured.  If
+        you would like to use this desktop please contact the system
+        administrator for {' '}<em>{clusterName}</em> and ask them to prepare
+        this desktop.
+      </div>
+    );
+  }
+
+  return {
+    body,
+    icon: 'danger',
+    header: 'Failed to launch desktop',
+  };
 }
 
 export default LaunchDesktopButton
